+++ example_runner.py
@@ -10,6 +10,7 @@ import example.house.house_plate
 import example.house.house_cut_away
 import example.house.stucco_brick_body
 import example.house.tudor_split_body
+import example.house.tudor_body
 
 #------------------------------
 ## Wall Examples