+++ src/cqfantasy/house/__init__.py
@@ -2,4 +2,5 @@ from .Body import Body
 from .Roof import Roof
 from .House import House
 from .StuccoBrickBody import StuccoBrickBody
+from .TudorBody import TudorBody
 from .TudorSplitBody import TudorSplitBody